refactor(test) : 프러덕션 코드 변경 적용
